First 10 values on DS1052E return trash.
On my DS1052E I need to throw the first 10
values that I get back from the scope away.
They are contain only trash.
If you realize that this behaviour is different
on a different scope, than please patch it
or at least notify me which scope it is and
what the behaviour should be like.